requirements.txt
chardet==5.2.0
Deprecated==1.2.14
img2pdf==0.5.1
lxml==5.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pikepdf==8.15.1
pillow==10.3.0
reportlab==4.2.0
wrapt==1.16.0
